mcp-dataverse-server/pyproject.toml
[project]
name = "mcp-dataverse-server"
version = "0.1.0"
description = "Local MCP server scaffold for Microsoft Dataverse Contacts"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "mcp>=0.0.0",
  "msal>=1.30",
  "python-dotenv>=1.0",
  "tzdata>=2024.1",
  "uvicorn>=0.27",
]


[project.scripts]
mcp-dataverse-server = "mcp_dataverse_server.__main__:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
